requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
jinja2==3.1.3
python-multipart==0.0.9
python-dotenv==1.0.1
